code/mqtt/requirements.txt
paho-mqtt==1.6.1
PyMySQL==1.1.0
cryptography